third_party/blink/web_tests/fast/canvas/color-space/canvas-colorspace-arguments.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>

var testScenarios = [
    ["Test default context creation parameters: srgb/unorm8",
        {},
        {colorSpace: "srgb", colorType: "unorm8"}],
    ["Test CanvasColorSpace value srgb",
        {colorSpace: "srgb"},
        {colorSpace: "srgb", colorType: "unorm8"}],
    ["Test CanvasColorSpace value rec2020, no pixel format, falls back to rec2020/unorm8",
        {colorSpace: "rec2020"},
        {colorSpace: "rec2020", colorType: "unorm8"}],
    ["Test CanvasColorSpace value p3, no pixel format, falls back to p3/unorm8",
        {colorSpace: "display-p3"},
        {colorSpace: "display-p3", colorType: "unorm8"}],

    ["Test CanvasPixelFormat value unorm8",
        {colorType: "unorm8"},
        {colorSpace: "srgb", colorType: "unorm8"}],
    ["Test CanvasPixelFormat value float16",
        {colorType: "float16"},
        {colorSpace: "srgb", colorType: "float16"}],

    ["Test supported color settings srgb/unorm8",
        {colorSpace: "srgb", colorType: "unorm8"},
        {colorSpace: "srgb", colorType: "unorm8"}],
    ["Test supported color settings srgb/float16",
        {colorSpace: "srgb", colorType: "float16"},
        {colorSpace: "srgb", colorType: "float16"}],

    ["Test supported color settings rec2020/unorm8, falls back to srgb/unorm8",
        {colorSpace: "rec2020", colorType: "unorm8"},
        {colorSpace: "rec2020", colorType: "unorm8"}],
    ["Test supported color settings rec2020/float16",
        {colorSpace: "rec2020", colorType: "float16"},
        {colorSpace: "rec2020", colorType: "float16"}],

    ["Test supported color settings p3/unorm8, falls back to srgb/unorm8",
        {colorSpace: "display-p3", colorType: "unorm8"},
        {colorSpace: "display-p3", colorType: "unorm8"}],
    ["Test supported color settings p3/float16",
        {colorSpace: "display-p3", colorType: "float16"},
        {colorSpace: "display-p3", colorType: "float16"}],
];

function runTestScenario(contextCreationParameters, expectedContextAttributes) {
    var canvas = document. createElement('canvas');
    var ctx = canvas.getContext('2d', contextCreationParameters);
    var contextAttributes = ctx.getContextAttributes();
    assert_equals(contextAttributes.colorSpace, expectedContextAttributes.colorSpace);
    assert_equals(contextAttributes.colorType, expectedContextAttributes.colorType);
}

generate_tests(runTestScenario, testScenarios);
</script>
